refactor: move task url generation to decoder and resolve router circular dependency
