+++ webui/react/src/utils/types.ts
@@ -1,4 +1,4 @@
-import { serverAddress } from 'routes/utils';
+// import { serverAddress } from 'routes/utils';
 import {
   AnyTask, Checkpoint, Command, CommandState, CommandType, ExperimentHyperParams,
   ExperimentItem, RawJson, RecentCommandTask, RecentExperimentTask, RecentTask, RunState, Step,
@@ -8,21 +8,6 @@ import { deletePathList, getPathList, isNumber, setPathList } from './data';
 import { getDuration } from './time';
 
 /* Conversions to Tasks */
-
-const commandToEventUrl = (command: Command): string | undefined => {
-  if (command.kind === CommandType.Notebook) return `/notebooks/${command.id}/events`;
-  if (command.kind === CommandType.Tensorboard) return `/tensorboard/${command.id}/events?tail=1`;
-  return undefined;
-};
-
-export const waitPageUrl = (command: Command): string | undefined => {
-  const eventUrl = commandToEventUrl(command);
-  const proxyUrl = command.serviceAddress;
-  if (!eventUrl || !proxyUrl) return;
-  const event = encodeURIComponent(eventUrl);
-  const jump = encodeURIComponent(proxyUrl);
-  return serverAddress(`/wait?event=${event}&jump=${jump}`);
-};
 
 export const commandToTask = (command: Command): RecentCommandTask => {
   // We expect the name to be in the form of 'Type (pet-name-generated)'.
@@ -38,7 +23,7 @@ export const commandToTask = (command: Command): RecentCommandTask => {
     startTime: command.registeredTime,
     state: command.state as CommandState,
     type: command.kind,
-    url: waitPageUrl(command),
+    url: command.url,
     username: command.user.username,
   };
   return task;
